Show w/kg J/kg estimates on Ride Summary
.. and fixes up the calculation too.
.. still need to fix up the LTM plot

+++ src/PDModel.h
@@ -140,7 +140,7 @@ class PDEstimate
 
         QDate from, to;
         QString model;
-        int WPrime,
+        double WPrime,
             CP,
             FTP,
             PMax;
